+++ src/jsx/index.jsx
@@ -14,12 +14,11 @@ import Start from './pages/start.jsx'
 import Edit from './pages/edit.jsx'
 import Publish from './pages/publish.jsx'
 import View from './pages/view.jsx'
-import Home from './pages/home.jsx'
 import SearchCodeAndComment from './pages/search-code-and-comment.jsx'
 import SearchComment from './pages/search-comment.jsx'
 
 
-const REDIRECT_URLS = ['/edit', '/publish', '/update', '/search_code_and_comment', '/search_comment']
+const REDIRECT_URLS = ['/publish', '/search_code_and_comment', '/search_comment']
 
 
 class CodeAndComment extends Component {
@@ -31,7 +30,7 @@ class CodeAndComment extends Component {
 
   changeRoute(event) {
     if (this.isFirst && REDIRECT_URLS.includes(event.url)) {
-      route('/home')
+      route('/edit')
     }
     this.isFirst = false
   }
@@ -40,13 +39,12 @@ class CodeAndComment extends Component {
     return (
       <Provider store={ store }>
         <Router history={ createHashHistory() } onChange={ this.changeRoute }>
-          <Edit path="/edit" />
           <Publish path="/publish" />
           <View path="/view" />
           <Start path="/start" />
           <SearchCodeAndComment path="/search_code_and_comment" />
           <SearchComment path="/search_comment" />
-          <Home default />
+          <Edit default />
         </Router>
       </Provider>
     )
