+++ src/scss/index.scss
@@ -2,7 +2,6 @@
 @import 'pages/edit';
 @import 'pages/view';
 @import 'pages/publish';
-@import 'pages/home';
 @import 'pages/search-code-and-comment';
 @import 'pages/search-comment';
