--- src/scss/pages/_home.scss
@@ -1,14 +1,0 @@
-@import '../variables';
-
-.cc-home {
-  @extend .cc-center;
-  width: fit#{"-"}content;
-  height: 100%;
-  .message {
-    padding-top: 16px;
-  }
-  .links {
-    font-size: 14px;
-    padding: 14px;
-  }
-}
