+++ test/actions/test_start.js
@@ -7,16 +7,6 @@ import actions from '../../src/jsx/actions/start.jsx'
 describe('actions/start', () => {
   function noop() {}
 
-  describe('home', () => {
-    it('moves /home', () => {
-      const route = spy()
-      const result = actions().home(null, null, route)
-      expect(route.calledOnce).to.be.true
-      expect(route.calledWith('/home')).to.be.true
-      expect(result).to.be.undefined
-    })
-  })
-
   describe('setLoading', () => {
     it('returns object which has loading key', () => {
       expect(actions().setLoading(null, true)).to.deep.equal({ loading: true })
